build TinyTeX from source, customize  package list
The packages standalone, varwidth and scontents are needed for the default
LaTeX preamble in Xournal++. The packages pgf, gfplots are used for
creating plots and graphics. The listings package is used for code
highlighting and the xargs package for defining LaTeX  commands
with multiple arguments. The other packages are those from TinyTeX.